Training results for smaller safe region (radius 1), larger controller loss (1e-4) are below
Epoch 12 training loss: 0.00028180998645317815
                     CLF origin: 1.8033215873307574e-05
           CLF safe region term: 4.6656509309819183e-07
                  (% satisfied): 99.99928507596069
         CLF unsafe region term: 0.0001524536721620116
                  (% satisfied): 99.82144732696145
               CLF descent term: 0.0033498903249039565
                  (% satisfied): 99.34857142857143
            CLF relaxation term: 0.0
                controller term: 3.083746311853124e-05
Epoch 12     test loss: 0.003551681241150905

Ntrain only 200K, so we see some generalization issues for the test set not having as small of a loss as the training set (Ntrain = 10K).

The main issue is that the descent term is not satisfied everywhere (second is there exists V < 1 in the unsafe region). We'll see how this plays out in simulation.